+++ ReleaseNotes/release_notes.md
@@ -5,6 +5,7 @@
   1. add generalized extreme value distribution: function and PDF
   1. impoetant improvements for classical reweigting 
   1. make more accurate "density" method for histograms  
+  1. add `table2` fuction for trees and datasets  
 
 ## Backward incompatible: